chore(eslint): Initialize configuration and enforce zero warnings
- Migrate configuration file from `eslint.config.js` to `eslint.config.ts`
- Install `jiti` as devDependency
- Add `--max-warnings=0` flag to `lint` script
- Add "node" to the `compilerOptions.types``
- Clean install `node_modules` and regenerate pnpm configuration
