+++ package.json
@@ -10,7 +10,7 @@
     "prepare": "svelte-kit sync || echo ''",
     "check": "wrangler types --check && svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
     "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
-    "lint": "prettier --check . && eslint .",
+    "lint": "prettier --check . && eslint . --max-warnings=0",
     "format": "prettier --write .",
     "test:unit": "vitest",
     "test": "npm run test:unit -- --run && npm run test:e2e",
@@ -36,6 +36,7 @@
     "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-svelte": "^3.19.0",
     "globals": "^17.6.0",
+    "jiti": "^2.7.0",
     "playwright": "^1.60.0",
     "prettier": "^3.8.3",
     "prettier-plugin-svelte": "^4.1.0",
