+++ pnpm-workspace.yaml
@@ -1,4 +1,4 @@
 allowBuilds:
-  workerd: true
-  sharp: true
   esbuild: true
+  sharp: true
+  workerd: true
