+++ tsconfig.json
@@ -11,6 +11,6 @@
     "sourceMap": true,
     "strict": true,
     "moduleResolution": "bundler",
-    "types": ["./worker-configuration.d.ts"]
+    "types": ["node", "./worker-configuration.d.ts"]
   }
 }
